+++ lib/pretender.rb
@@ -6,7 +6,11 @@ module Pretender
 
   def impersonates(scope = :user, opts = {})
     impersonated_method = opts[:method] || :"current_#{scope}"
-    impersonate_with = opts[:with] || proc { |id| scope.to_s.classify.constantize.where(:id => id).first }
+    impersonate_with = opts[:with] || proc { |id|
+      klass = scope.to_s.classify.constantize
+      primary_key = klass.respond_to?(:primary_key) ? klass.primary_key : :id
+      klass.find_by(primary_key => id)
+    }
     true_method = :"true_#{scope}"
     session_key = :"impersonated_#{scope}_id"
     impersonated_var = :"@impersonated_#{scope}"